+++ packages/core/index.ts
@@ -86,7 +86,7 @@ export * from './useSpeechRecognition'
 export * from './useSpeechSynthesis'
 export * from './useStorage'
 export * from './useStorageAsync'
-export * from './useStyle'
+export * from './useStyleTag'
 export * from './useSwipe'
 export * from './useTemplateRefsList'
 export * from './useTextSelection'
